+++ src/components/DndMultiProvider/DndMultiProvider.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from 'react';
+import { DndProvider } from 'react-dnd';
+import MultiBackend from 'react-dnd-multi-backend';
+import HTML5toTouch from 'react-dnd-multi-backend/dist/esm/HTML5toTouch';
+
+export const DndMultiProvider: FC = ({ children }) => {
+  return (
+    <DndProvider backend={MultiBackend} options={HTML5toTouch}>
+      {children}
+    </DndProvider>
+  );
+};