+++ src/components/DndMultiProvider/index.ts
@@ -1,0 +1,1 @@
+export { DndMultiProvider as default } from './DndMultiProvider';