+++ src/components/DragLayer/DragLayer.stories.tsx
@@ -1,6 +1,5 @@
 import React, { CSSProperties } from 'react';
-import { DndProvider } from 'react-dnd';
-import Backend from 'react-dnd-html5-backend';
+import DndMultiProvider from '../DndMultiProvider';
 import Draggable from '../Draggable';
 import DragLayer from '../DragLayer';
 import docs from './README.md';
@@ -19,18 +18,18 @@ const commonStyle: CSSProperties = {
 };
 
 export const basic = () => (
-  <DndProvider backend={Backend}>
+  <DndMultiProvider>
     <Draggable index={0} hideDragPreview={true}>
       <div style={{ ...commonStyle, border: '1px solid red' }}>This div is draggable!</div>
     </Draggable>
     <DragLayer>
       <div style={{ ...commonStyle, border: '1px solid blue', opacity: 0.9 }}>Custom preview!</div>
     </DragLayer>
-  </DndProvider>
+  </DndMultiProvider>
 );
 
 export const withCustomTranslate = () => (
-  <DndProvider backend={Backend}>
+  <DndMultiProvider>
     <Draggable index={0} hideDragPreview={true}>
       <div style={{ ...commonStyle, border: '1px solid red' }}>This div is draggable!</div>
     </Draggable>
@@ -43,5 +42,5 @@ export const withCustomTranslate = () => (
     >
       <div style={{ ...commonStyle, border: '1px solid blue', opacity: 0.9 }}>Custom preview!</div>
     </DragLayer>
-  </DndProvider>
+  </DndMultiProvider>
 );