+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - No longer takes onRemove and onRotate
   - `buttonProps` is array of objects with properties `onClick` and `children`
   - Allows for flexibility when defining button on `Thumbnail`
+- [DndMultiProvider] Supports touch and mouse events (#24) @liamross
 
 ## [0.0.0-beta.4] - 2020-01-09